refactor(drc): Rename overlapping pad utility and update imports
- Renamed the `checkOverlappingPads` utility file to follow a consistent naming convention.
- Updated import paths in the Board component and related tests to reflect the new file name.
- Removed the old `checkOverlappingPads.ts` file as part of the refactor.
